+++ LaserAbsorptionSolver.h
@@ -241,7 +241,7 @@ private:
                                                SpaceVariable3D *V = NULL); //if NULL, use stored temperature
   
   inline double GetAbsorptionCoefficient(double T, int id) { //T must be in Kelvin
-    return id<absorption.size() ? 
+    return id<(int)absorption.size() ? 
              std::get<0>(absorption[id])*(T - std::get<1>(absorption[id])) + std::get<2>(absorption[id])
            : 0.0; //could get here if there are "inactive" nodes
   }
